added images and practicing html
